RST-1778: Summary Page Revamp (#229)
Updates strings and order of elements on the Confirmation of Supplied Details page.

Data was showing many users were dropping off on this page, presumably because they thought they had completed the journey. These changes aim to make it clearer that the journey has not been completed, thus decreasing exits on this page and increasing completions.